Put the std::string entity_id and Entity *entity system in a separate HasEntity class
