+++ vagabond/core/meson.build
@@ -139,6 +139,8 @@ cfile,
 'GeometryTable.cpp',
 'Grapher.cpp',
 'HasBondstraints.cpp',
+'HasEntity.h',
+'HasEntity.cpp',
 'HelpKnot.cpp',
 'HelpKnot.h',
 'HyperValue.cpp',
